first commit to blog app

+++ blogs/views.py
@@ -2,8 +2,10 @@ from django.shortcuts import render,redirect
 
 from . models import BlogPost
 from . forms import BlogForm
+from django.contrib.auth.decorators import login_required
+from django.http import Http404
 
-# Create your views here.
+# Create your views here.s
 def index(request):
 
     blogposts = BlogPost.objects.order_by('date_added')
@@ -11,7 +13,7 @@ def index(request):
     return render(request, "blogs/index.html", context)
 
 
-
+@login_required
 def new_blog(request):
     """Add new Blog"""
 
@@ -21,6 +23,8 @@ def new_blog(request):
         form = BlogForm(data=request.POST)
 
         if form.is_valid():
+            new_blog = form.save(commit=False)
+            new_blog.owner = request.user
             form.save()
             return redirect("blogs:index")
     
@@ -28,12 +32,13 @@ def new_blog(request):
     return render(request, "blogs/new_blog.html", context)
 
 
-
+@login_required
 def edit_blog(request, blog_id):
     """edit existing blog"""
 
     blogpost = BlogPost.objects.get(id=blog_id)
-
+    #check for post owner raise 404 if not the owner.
+    check_blog_owner(blogpost.owner, request.user)
 
     if request.method != "POST":
         form = BlogForm(instance=blogpost)
@@ -47,11 +52,14 @@ def edit_blog(request, blog_id):
     context = {'blogpost': blogpost, 'form': form }
     return render(request, "blogs/edit_blog.html", context)
 
-
+@login_required
 def delete_blog(request, blog_id):
     """delete existing blog"""
 
     blogpost = BlogPost.objects.get(id=blog_id)
+    
+    #check for post owner raise 404 if not the owner.
+    check_blog_owner(blogpost.owner, request.user)
     
     if request.method == "POST":
         blogpost.delete()
@@ -59,3 +67,9 @@ def delete_blog(request, blog_id):
         
     context = {'blogpost': blogpost}
     return render(request, "blogs/delete_blog.html", context)
+
+
+def check_blog_owner(owner, request):
+
+    if owner != request:
+        raise Http404
